+++ src/lib/auth/get-current-user.ts
@@ -1,15 +1,15 @@
-import axios, { AxiosError } from 'axios';
+import axios from 'axios';
 import type { User } from '$lib/types/user';
 import Maybe from '$lib/types/maybe';
+import {handleErrors} from "$lib/utils/requests";
 
-export async function getCurrentUser(): Promise<Maybe<User, string>> {
+export async function getCurrentUser(): Promise<Maybe<User, string[]>> {
 	try {
 		const user = await axios.get<User>('http://localhost:3000/auth/profile', {
 			withCredentials: true
 		});
 		return Maybe.ok(user.data);
 	} catch (e) {
-		const error: string = (e as AxiosError).response?.data?.message;
-		return Maybe.error(error);
+		return handleErrors<User>(e)
 	}
 }
