+++ src/routes/dashboard/UserInfo.svelte
@@ -1,4 +1,4 @@
-<script lang="ts">
+<script>
     export let userData = {}
 </script>
